--- a/alias_test.go
+++ b/alias_test.go
@@ -9,6 +9,8 @@ import (
 	"math/rand"
 	"reflect"
 	"testing"
+
+	stat "github.com/ematvey/gostat"
 )
 
 const distributionCount = 1000000
@@ -47,6 +49,65 @@ func TestDistribution(t *testing.T) {
 	testDistribution(t, []float64{9, 8, 1, 4, 2}, 5)
 	testDistribution(t, []float64{1000, 1, 3, 10}, 39)
 	testDistribution(t, []float64{1000, 1, 3, 10}, 61)
+}
+
+func TestTail(t *testing.T) {
+	const size = 33294320
+	const half = size / 2
+	const tries = 1000000
+	const alpha = 0.05
+	dist := make([]float64, size)
+	for i := range dist {
+		dist[i] = 1
+	}
+	a, err := New(dist)
+	if err != nil {
+		t.Fatalf("Got an error during creation:", err)
+	}
+	rng := rand.New(rand.NewSource(42))
+	var k int64 // [0,half) (one half)
+	for i := 0; i < tries; i++ {
+		if a.Gen(rng) < half {
+			k++
+		}
+	}
+	// Expected probability of getting k <= k_observed if p == 0.5.
+	p := stat.Binomial_CDF_At(0.5, tries, k)
+	if p < alpha/2 || p > (1-alpha/2) {
+		t.Errorf("The distribution is biased. %d of %d results were in the first half. Binomial_CDF = %f", k, tries, p)
+	}
+}
+
+func TestBalanceInsideBucket(t *testing.T) {
+	const size = 33294320
+	//const size = 8
+	const half = size / 2
+	const tries = 1000000
+	const alpha = 0.05
+	dist := make([]float64, size)
+	for i := range dist {
+		if i < half {
+			dist[i] = 1
+		} else {
+			dist[i] = 3
+		}
+	}
+	a, err := New(dist)
+	if err != nil {
+		t.Fatalf("Got an error during creation:", err)
+	}
+	rng := rand.New(rand.NewSource(421))
+	var k int64 // [0,half) (one half)
+	for i := 0; i < tries; i++ {
+		if a.Gen(rng) < half {
+			k++
+		}
+	}
+	// Expected probability of getting k <= k_observed if p == 0.5.
+	p := stat.Binomial_CDF_At(0.25, tries, k)
+	if p < alpha/2 || p > (1-alpha/2) {
+		t.Errorf("The distribution is biased. %d of %d results were in the first half. Binomial_CDF = %f", k, tries, p)
+	}
 }
 
 func TestMarshalBinary(t *testing.T) {