--- a/alias_test.go
+++ b/alias_test.go
@@ -5,6 +5,7 @@
 package alias
 
 import (
+	"fmt"
 	"math"
 	"math/rand"
 	"reflect"
@@ -111,30 +112,109 @@ func TestBalanceInsideBucket(t *testing.T) {
 }
 
 func TestMarshalBinary(t *testing.T) {
-	distributions := [][]float64{
-		{1},
-		{1, 1},
-		{1, 2, 3, 4, 5, 6, 7, 8, 9, 1000},
-	}
-	for _, distribution := range distributions {
-		a, err := New(distribution)
+	makeFloat := func(p []float64) *Alias {
+		a, err := New(p)
 		if err != nil {
 			t.Fatalf("Couldn't create alias: %v", err)
 		}
-
+		return a
+	}
+	makeInt := func(p []int32) *Alias {
+		a, err := NewInt(p)
+		if err != nil {
+			t.Fatalf("Couldn't create alias: %v", err)
+		}
+		return a
+	}
+	aliases := []*Alias{
+		makeFloat([]float64{1}),
+		makeFloat([]float64{1, 1}),
+		makeFloat([]float64{1, 2, 3, 4, 5, 6, 7, 8, 9, 1000}),
+		makeInt([]int32{1}),
+		makeInt([]int32{1, 1}),
+		makeInt([]int32{1, 2, 3, 4, 5, 6, 7, 8, 9, 1000}),
+	}
+	for i, a := range aliases {
 		data, err := a.MarshalBinary()
 		if err != nil {
-			t.Fatalf("Couldn't MarshalBinary: %v", err)
+			t.Errorf("Couldn't MarshalBinary: %v", err)
 		}
 
 		a2 := &Alias{}
 		err = a2.UnmarshalBinary(data)
 		if err != nil {
-			t.Fatalf("Couldn't UnmarshalBinary: %v", err)
+			t.Errorf("Couldn't UnmarshalBinary: %v", err)
 		}
 
 		if !reflect.DeepEqual(a, a2) {
-			t.Fatalf("Unmarshalled version was not the same as original")
+			fmt.Println(a, a2)
+			t.Errorf("case %d: Unmarshalled version %v was not the same as original %v", i, a2, a)
 		}
+	}
+}
+
+func testIntDistribution(t *testing.T, dist []int32, seed int64) {
+	sum := uint64(0)
+	for i := 0; i < len(dist); i++ {
+		sum += uint64(dist[i])
+	}
+
+	a, err := NewInt(dist)
+	if err != nil {
+		t.Error("Got an error during creation:", err)
+		return
+	}
+
+	rng := rand.New(rand.NewSource(seed))
+
+	counts := make([]int64, len(dist))
+	for i := 0; i < distributionCount; i++ {
+		counts[a.Gen(rng)]++
+	}
+
+	for i := 0; i < len(dist); i++ {
+		p := float64(counts[i]) / distributionCount
+		if math.Abs(p-float64(dist[i])/float64(sum)) > errorBound {
+			t.Error("Distribution did not match, seed", seed, "- got ", p, "expected", float64(dist[i])/float64(sum))
+		}
+	}
+}
+
+func TestIntDistribution(t *testing.T) {
+	testIntDistribution(t, []int32{1, 1}, 1)
+	testIntDistribution(t, []int32{1, 2, 3}, 2)
+	testIntDistribution(t, []int32{9, 8, 1, 4, 2}, 5)
+	testIntDistribution(t, []int32{1000, 1, 3, 10}, 39)
+	testIntDistribution(t, []int32{1000, 1, 3, 10}, 61)
+}
+
+func TestIntBalanceInsideBucket(t *testing.T) {
+	const size = 33294320
+	const half = size / 2
+	const tries = 1000000
+	const alpha = 0.05
+	dist := make([]int32, size)
+	for i := range dist {
+		if i < half {
+			dist[i] = 1
+		} else {
+			dist[i] = 3
+		}
+	}
+	a, err := NewInt(dist)
+	if err != nil {
+		t.Fatalf("Got an error during creation:", err)
+	}
+	rng := rand.New(rand.NewSource(421))
+	var k int64 // [0,half) (one half)
+	for i := 0; i < tries; i++ {
+		if a.Gen(rng) < half {
+			k++
+		}
+	}
+	// Expected probability of getting k <= k_observed if p == 0.5.
+	p := stat.Binomial_CDF_At(0.25, tries, k)
+	if p < alpha/2 || p > (1-alpha/2) {
+		t.Errorf("The distribution is biased. %d of %d results were in the first half. Binomial_CDF = %f", k, tries, p)
 	}
 }