There were two issues.

 * Bias of distribution of remainders of Int31.
   Demonstrated in TestTail. Items from the first half of the table
   get one addition remainder. Can be fixed without overhead
   by rethrowing RNG in that rare case when its value goes belongs
   to that tail.

 * The random number used to choose bucket is not independent from
   the random number used to decide whether to switch to alias.
   Demonstrated in TestBalanceInsideBucket. I don't know how to
   fix without overhead of second call to RNG.

Both issues were resolved in this commit. Statistical tests added.

Impact on performance:

Before:

```
BenchmarkGen5-8                 100000000               24.1 ns/op
BenchmarkGen50-8                50000000                28.7 ns/op
BenchmarkGen500-8               50000000                27.2 ns/op
BenchmarkGen5000-8              50000000                27.9 ns/op
BenchmarkGen50000-8             50000000                31.3 ns/op
```

After:

```
BenchmarkGen5-8                 50000000                38.8 ns/op
BenchmarkGen50-8                30000000                53.0 ns/op
BenchmarkGen500-8               30000000                40.9 ns/op
BenchmarkGen5000-8              30000000                41.5 ns/op
BenchmarkGen50000-8             30000000                43.9 ns/op
```