@starius 2017-03-13
I merged these commits to master branch of my fork: https://github.com/starius/alias/